Added User superclass
Created a new abstract class named `User` (package `uom.backend.physioassistant.models.users`).
This is a superclass for other classes that will represent different types of users such as Doctor, Admin and Patient.
It has a single field for the password and is annotated with lombok annotations to generate getters, setters, and constructors.