Added Admin class to represent Hellenic Organization of Physiotherapists (ΠΣΦ)
Created a new class named `Admin` (package `uom.backend.physioassistant.models.users`).
This class represents the Hellenic Organization of Physiotherapists (ΠΣΦ) who will be using the application.
It has fields such as username and password, and inherits the password field from the User superclass. The class is annotated with the necessary annotations for persistence and lombok annotations for generating getters, setters, and constructors.