Added Doctor class to represent Physiotherapists
Created a new class named `Doctor` (package `uom.backend.physioassistant.models.users`).
This class represents the physiotherapists who will be using the application.
It has fields such as afm, name, and address, and inherits the password field from the User superclass. The class is annotated with the necessary annotations for persistence and lombok annotations for generating getters, setters, and constructors.