Added error handling to controllers
Added error handling to the controllers, in order to provide more  specific responses to the client when certain errors occur in the service layer.
With this implementation, if the service layer throws specific errors (eg EntityNotFound), the controller will catch them and return the corresponding response (eg 404) to the client.